requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
APScheduler==3.6.3
async-timeout==4.0.2
attrs==22.2.0
blobfile==2.0.1
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
docker-pycreds==0.4.0
et-xmlfile==1.1.0
filelock==3.9.0
fonttools==4.38.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.30
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.6.3
multidict==6.0.4
numpy==1.24.1
openai==0.26.0
openpyxl==3.0.10
packaging==23.0
pandas==1.5.2
pandas-stubs==1.5.2.230105
pathtools==0.1.2
Pillow==9.4.0
plotly==5.12.0
protobuf==4.21.12
psutil==5.9.4
pycryptodomex==3.16.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-telegram-bot==13.7
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rfc3986==1.5.0
scikit-learn==1.2.0
scipy==1.10.0
sentry-sdk==1.13.0
setproctitle==1.3.2
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
sniffio==1.3.0
tenacity==8.1.0
threadpoolctl==3.1.0
tiktoken==0.1.2
tornado==6.2
tqdm==4.64.1
types-pytz==2022.7.0.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
wandb==0.13.9
yarl==1.8.2
